:bug: Fix Bug Client/Server Component with next-translate for Contact Component
