+++ src/components/contact/contactForm/ContactForm.module.scss
@@ -1,0 +1,109 @@
+@import 'variables';
+
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 2rem;
+  margin: 2rem 0;
+  padding: 3rem;
+  min-height: 388px;
+  background: linear-gradient($bgGradiantDark, $bgGradiantLight);
+  border-radius: 0.4rem;
+  .message {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+    line-height: 1.8;
+    font-weight: 300;
+  }
+  .container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    justify-content: space-between;
+    gap: 3rem;
+    max-width: 1000px;
+    width: 100%;
+    .column {
+      display: flex;
+      flex-direction: column;
+      .input,
+      .textarea {
+        width: 100%;
+        margin-bottom: 1rem;
+        padding: 0.5rem;
+        background-color: $violetContact;
+        border: 1px solid $violetContact;
+        border-radius: 0.4rem;
+        color: $violetLight;
+        font-size: 0.9rem;
+        font-weight: 300;
+        transition: border-color 0.3s ease-in-out;
+        &::placeholder {
+          color: $violetLight;
+        }
+        &:focus {
+          border: 1px solid $violetLight;
+        }
+      }
+      .textarea {
+        resize: none;
+        outline: none;
+        height: 100%;
+      }
+      .error {
+        display: flex;
+        align-items: center;
+        gap: 0.5rem;
+        font-size: 0.8rem;
+        font-weight: 300;
+        padding-bottom: 1rem;
+        padding-left: 0.5rem;
+      }
+    }
+  }
+  .button {
+    width: 160px;
+    height: 40px;
+    background: linear-gradient(45deg, $blue, $violet 30%, $violetPink);
+    border-radius: 3rem;
+    box-shadow: 0 2px 4px 0 $shadow;
+    color: $violetLight;
+    font-size: 0.9rem;
+    text-align: center;
+    text-transform: uppercase;
+    transition: all 0.3s ease-in-out;
+    &:hover {
+      transform: scale(1.05);
+    }
+  }
+}
+
+//Responsive screen < 999px
+@media (max-width: 999px) {
+  .form {
+    .container {
+      flex-direction: column;
+      grid-template-columns: 1fr;
+      gap: 0rem;
+      .column {
+        resize: none;
+        outline: none;
+        .textarea {
+          resize: none;
+          outline: none;
+          height: 200px;
+        }
+      }
+    }
+  }
+}
+
+//Responsive screen < 768px
+@media (max-width: 768px) {
+  .form {
+    padding: 1.5em;
+  }
+}
